Change return type for diff to Double as it can be bigger than a regular Int

+++ src/main/scala/moment/Date.scala
@@ -38,7 +38,7 @@ trait Date extends js.Object with Getters with Setters[Date] {
   def isoWeekday(): Int = js.native
   def isoWeekday(newDay: Int): Date = js.native
   def diff(date: Date): Int = js.native
-  def diff(date: Date, unit: String): Int = js.native
+  def diff(date: Date, unit: String): Double = js.native
   def diff(date: Date, unit: String, asFloat: Boolean): Double = js.native
   def local(): Date = js.native
   def utc(): Date = js.native
